Extract lightweight model metadata and installed lookup
