+++ Sources/MereRunCore/ModelManagementExports.swift
@@ -1,0 +1,2 @@
+// Preserve existing Core imports while model-management types have a lightweight owner.
+@_exported import MereRunModelKit
