+++ Sources/MereRunCore/README.md
@@ -1,7 +1,8 @@
 # MereRunCore
 
-Shared model resolution, manifests, generation protocols, and native runtime
-families.
+Shared runtime validation, catalog assembly, generation protocols, and native
+runtime families. Model metadata and installed lookup live in `MereRunModelKit`;
+Core re-exports those types for existing callers.
 
 - `Generation.swift`: image and chat request/response contracts.
 - `ImageGenerationOptions.swift` and `ImageGenerationPlan.swift`: typed image
@@ -13,7 +14,8 @@ families.
 - `ManagedAdapterArgumentResolver.swift`: installed adapter lookup and base-model
   compatibility shared by image, text, video, and evaluation adapters.
 - `ManagedModel*.swift`: public managed-model catalog and install metadata.
-- `MereRunModel*.swift`: manifests, paths, and validation.
+- `ModelResolver.swift`: adapts the catalog and family validators to ModelKit lookup.
+- `MereRunModelManifest+Templates.swift`: runtime-dependent manifest templates.
 - Runtime family directories own model-specific loading, inference, and decode
   paths.
